Update GetUserSchema to include auctions created.
